Velocity gridded timeseries product code
An attempt to create a velocity gridded product using the groupby method, as used to produce the other gridded products. Gridded VCUR, UCUR and WCUR are provided, along with metadata.

The code could probably do with improving - I tried to run the loop in parallel with no luck. Although for one year's worth of velocity data it takes about 4 mins on my laptop without remote access to a server so not too bad.

Attributes are mostly correct, but a few attributes might need tweaking or new ones added. Also keep in mind that the method of adding attributes might not be satisfactory for the AODN. For example, I have copied the abstract, title and lineage attributes into the script rather than relying on the json template.

This product is probably good enough for NSW-IMOS purposes, but we may develop different versions, or improve the code in future. If so, I can create a new pull request.